base: Introduce Accidental Touch
Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

@idoybh: Add settings value to backup array

Change-Id: I21a8f12364c0d0072fcf6b7bc61e315520a6b047
Signed-off-by: Zaviyar786 <beingme639@gmail.com>